day02: refactor sub into its own thing
