+++ day02/src/sub/movement.rs
@@ -1,0 +1,5 @@
+pub trait SubPosition {
+    fn forward(&mut self, by: i32);
+    fn down(&mut self, by: i32);
+    fn up(&mut self, by: i32);
+}
